SQUASHME! [VPG]: drm/i915: Improved scheduler trace output
Seqno values are no longer valid at all times due to being lazily
allocated. Instead, the request structure has a 'uniq' field with is
valid at all times. This patch updates the scheduler's trace output
strings to include the uniq field so that they are actually useful.

This patch could be squashed in to the one that originally added the
scheduler trace entries:
  commit c64b5c1f236ee9c9362f06a675b4bc906ad3cb04
  Author: John Harrison <John.C.Harrison@Intel.com>
  Date:   Thu Apr 10 10:58:56 2014 +0100
  Subject: FOR_UPSTREAM [VPG]: drm/i915: Added trace points to scheduler
  Change-id: I9886390cfc7897bc1faf50a104bc651d8baed8a5

Change-Id: I7334934f3dc20f938355ba95758ce631d280b724
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-45002
Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-on: https://android.intel.com:443/421475